refactor: move package and delete javadocs
