--- docs/allpackages-index.html
@@ -1,63 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-<!-- Generated by javadoc (21) on Mon Mar 11 15:17:09 WIB 2024 -->
-<title>All Packages</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-03-11">
-<meta name="description" content="package index">
-<meta name="generator" content="javadoc/AllPackagesIndexWriter">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="script-dir/jquery-ui.min.css" title="Style">
-<script type="text/javascript" src="script.js"></script>
-<script type="text/javascript" src="script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="script-dir/jquery-ui.min.js"></script>
-</head>
-<body class="all-packages-index-page">
-<script type="text/javascript">var pathtoroot = "./";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<div class="flex-box">
-<header role="banner" class="flex-header">
-<nav role="navigation">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
-<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
-<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="com/elangsegara/package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="overview-tree.html">Tree</a></li>
-<li><a href="index-files/index-1.html">Index</a></li>
-<li><a href="help-doc.html#all-packages">Help</a></li>
-</ul>
-</div>
-<div class="sub-nav">
-<div id="navbar-sub-list"></div>
-<div class="nav-list-search"><a href="search.html">SEARCH</a>
-<input type="text" id="search-input" disabled placeholder="Search">
-<input type="reset" id="reset-button" disabled value="reset">
-</div>
-</div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<span class="skip-nav" id="skip-navbar-top"></span></nav>
-</header>
-<div class="flex-content">
-<main role="main">
-<div class="header">
-<h1 title="All&amp;nbsp;Packages" class="title">All&nbsp;Packages</h1>
-</div>
-<div class="caption"><span>Package Summary</span></div>
-<div class="summary-table two-column-summary">
-<div class="table-header col-first">Package</div>
-<div class="table-header col-last">Description</div>
-<div class="col-first even-row-color"><a href="com/elangsegara/package-summary.html">com.elangsegara</a></div>
-<div class="col-last even-row-color">&nbsp;</div>
-</div>
-</main>
-</div>
-</div>
-</body>
-</html>
