--- docs/com/elangsegara/package-tree.html
@@ -1,77 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-<!-- Generated by javadoc (21) on Mon Mar 11 15:17:09 WIB 2024 -->
-<title>com.elangsegara Class Hierarchy</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-03-11">
-<meta name="description" content="tree: package: com.elangsegara">
-<meta name="generator" content="javadoc/PackageTreeWriter">
-<link rel="stylesheet" type="text/css" href="../../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../../script-dir/jquery-ui.min.css" title="Style">
-<script type="text/javascript" src="../../script.js"></script>
-<script type="text/javascript" src="../../script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="../../script-dir/jquery-ui.min.js"></script>
-</head>
-<body class="package-tree-page">
-<script type="text/javascript">var pathtoroot = "../../";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<div class="flex-box">
-<header role="banner" class="flex-header">
-<nav role="navigation">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
-<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
-<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="package-summary.html">Package</a></li>
-<li>Class</li>
-<li class="nav-bar-cell1-rev">Tree</li>
-<li><a href="../../index-files/index-1.html">Index</a></li>
-<li><a href="../../help-doc.html#tree">Help</a></li>
-</ul>
-</div>
-<div class="sub-nav">
-<div id="navbar-sub-list"></div>
-<div class="nav-list-search"><a href="../../search.html">SEARCH</a>
-<input type="text" id="search-input" disabled placeholder="Search">
-<input type="reset" id="reset-button" disabled value="reset">
-</div>
-</div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<span class="skip-nav" id="skip-navbar-top"></span></nav>
-</header>
-<div class="flex-content">
-<main role="main">
-<div class="header">
-<h1 class="title">Hierarchy For Package com.elangsegara</h1>
-</div>
-<section class="hierarchy">
-<h2 title="Class Hierarchy">Class Hierarchy</h2>
-<ul>
-<li class="circle">java.lang.<a href="https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/lang/Object.html" class="type-name-link external-link" title="class or interface in java.lang">Object</a>
-<ul>
-<li class="circle">org.gradle.api.internal.AbstractTask (implements org.gradle.api.internal.DynamicObjectAware, org.gradle.api.internal.TaskInternal)
-<ul>
-<li class="circle">org.gradle.api.DefaultTask (implements org.gradle.api.Task)
-<ul>
-<li class="circle">com.elangsegara.<a href="XrayTask.html" class="type-name-link" title="class in com.elangsegara">XrayTask</a></li>
-</ul>
-</li>
-</ul>
-</li>
-<li class="circle">com.elangsegara.<a href="XrayExtension.html" class="type-name-link" title="class in com.elangsegara">XrayExtension</a></li>
-<li class="circle">com.elangsegara.<a href="XrayMain.html" class="type-name-link" title="class in com.elangsegara">XrayMain</a></li>
-<li class="circle">com.elangsegara.<a href="XrayPlugin.html" class="type-name-link" title="class in com.elangsegara">XrayPlugin</a> (implements org.gradle.api.Plugin&lt;T&gt;)</li>
-</ul>
-</li>
-</ul>
-</section>
-</main>
-</div>
-</div>
-</body>
-</html>
