--- docs/index-files/index-7.html
@@ -1,84 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-<!-- Generated by javadoc (21) on Mon Mar 11 15:17:09 WIB 2024 -->
-<title>X-Index</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-03-11">
-<meta name="description" content="index: X">
-<meta name="generator" content="javadoc/IndexWriter">
-<link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
-<link rel="stylesheet" type="text/css" href="../script-dir/jquery-ui.min.css" title="Style">
-<script type="text/javascript" src="../script.js"></script>
-<script type="text/javascript" src="../script-dir/jquery-3.6.1.min.js"></script>
-<script type="text/javascript" src="../script-dir/jquery-ui.min.js"></script>
-</head>
-<body class="index-page">
-<script type="text/javascript">var pathtoroot = "../";
-loadScripts(document, 'script');</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<div class="flex-box">
-<header role="banner" class="flex-header">
-<nav role="navigation">
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="top-nav" id="navbar-top"><button id="navbar-toggle-button" aria-controls="navbar-top" aria-expanded="false" aria-label="Toggle navigation links"><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span><span class="nav-bar-toggle-icon">&nbsp;</span></button>
-<div class="skip-nav"><a href="#skip-navbar-top" title="Skip navigation links">Skip navigation links</a></div>
-<ul id="navbar-top-firstrow" class="nav-list" title="Navigation">
-<li><a href="../com/elangsegara/package-summary.html">Package</a></li>
-<li>Class</li>
-<li><a href="../overview-tree.html">Tree</a></li>
-<li class="nav-bar-cell1-rev">Index</li>
-<li><a href="../help-doc.html#index">Help</a></li>
-</ul>
-</div>
-<div class="sub-nav">
-<div id="navbar-sub-list"></div>
-<div class="nav-list-search"><a href="../search.html">SEARCH</a>
-<input type="text" id="search-input" disabled placeholder="Search">
-<input type="reset" id="reset-button" disabled value="reset">
-</div>
-</div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<span class="skip-nav" id="skip-navbar-top"></span></nav>
-</header>
-<div class="flex-content">
-<main role="main">
-<div class="header">
-<h1>Index</h1>
-</div>
-<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">G</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">R</a>&nbsp;<a href="index-6.html">S</a>&nbsp;<a href="index-7.html">X</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
-<h2 class="title" id="I:X">X</h2>
-<dl class="index">
-<dt><a href="../com/elangsegara/XrayExtension.html" class="type-name-link" title="class in com.elangsegara">XrayExtension</a> - Class in <a href="../com/elangsegara/package-summary.html">com.elangsegara</a></dt>
-<dd>
-<div class="block">Extension class for configuring the XRAY Gradle Plugin.</div>
-</dd>
-<dt><a href="../com/elangsegara/XrayExtension.html#%3Cinit%3E()" class="member-name-link">XrayExtension()</a> - Constructor for class com.elangsegara.<a href="../com/elangsegara/XrayExtension.html" title="class in com.elangsegara">XrayExtension</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../com/elangsegara/XrayMain.html" class="type-name-link" title="class in com.elangsegara">XrayMain</a> - Class in <a href="../com/elangsegara/package-summary.html">com.elangsegara</a></dt>
-<dd>
-<div class="block">Main class for the XRAY utility.</div>
-</dd>
-<dt><a href="../com/elangsegara/XrayMain.html#%3Cinit%3E()" class="member-name-link">XrayMain()</a> - Constructor for class com.elangsegara.<a href="../com/elangsegara/XrayMain.html" title="class in com.elangsegara">XrayMain</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../com/elangsegara/XrayPlugin.html" class="type-name-link" title="class in com.elangsegara">XrayPlugin</a> - Class in <a href="../com/elangsegara/package-summary.html">com.elangsegara</a></dt>
-<dd>
-<div class="block">The main plugin class for the XRAY Gradle Plugin.</div>
-</dd>
-<dt><a href="../com/elangsegara/XrayPlugin.html#%3Cinit%3E()" class="member-name-link">XrayPlugin()</a> - Constructor for class com.elangsegara.<a href="../com/elangsegara/XrayPlugin.html" title="class in com.elangsegara">XrayPlugin</a></dt>
-<dd>&nbsp;</dd>
-<dt><a href="../com/elangsegara/XrayTask.html" class="type-name-link" title="class in com.elangsegara">XrayTask</a> - Class in <a href="../com/elangsegara/package-summary.html">com.elangsegara</a></dt>
-<dd>
-<div class="block">Task class for executing the XRAY Gradle Plugin.</div>
-</dd>
-<dt><a href="../com/elangsegara/XrayTask.html#%3Cinit%3E()" class="member-name-link">XrayTask()</a> - Constructor for class com.elangsegara.<a href="../com/elangsegara/XrayTask.html" title="class in com.elangsegara">XrayTask</a></dt>
-<dd>&nbsp;</dd>
-</dl>
-<a href="index-1.html">A</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">G</a>&nbsp;<a href="index-4.html">M</a>&nbsp;<a href="index-5.html">R</a>&nbsp;<a href="index-6.html">S</a>&nbsp;<a href="index-7.html">X</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes&nbsp;and&nbsp;Interfaces</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
-</div>
-</div>
-</body>
-</html>
