--- docs/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE HTML>
-<html lang="en">
-<head>
-<!-- Generated by javadoc (21) on Mon Mar 11 15:17:09 WIB 2024 -->
-<title>Generated Documentation (Untitled)</title>
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2024-03-11">
-<meta name="description" content="index redirect">
-<meta name="generator" content="javadoc/IndexRedirectWriter">
-<link rel="canonical" href="com/elangsegara/package-summary.html">
-<link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
-<script type="text/javascript">window.location.replace('com/elangsegara/package-summary.html')</script>
-<noscript>
-<meta http-equiv="Refresh" content="0;com/elangsegara/package-summary.html">
-</noscript>
-</head>
-<body class="index-redirect-page">
-<main role="main">
-<noscript>
-<p>JavaScript is disabled on your browser.</p>
-</noscript>
-<p><a href="com/elangsegara/package-summary.html">com/elangsegara/package-summary.html</a></p>
-</main>
-</body>
-</html>
