--- docs/member-search-index.js
@@ -1,1 +1,0 @@
-memberSearchIndex = [{"p":"com.elangsegara","c":"XrayPlugin","l":"apply(Project)","u":"apply(org.gradle.api.Project)"},{"p":"com.elangsegara","c":"XrayExtension","l":"getAction()"},{"p":"com.elangsegara","c":"XrayExtension","l":"getScenario()"},{"p":"com.elangsegara","c":"XrayMain","l":"main(String[])","u":"main(java.lang.String[])"},{"p":"com.elangsegara","c":"XrayTask","l":"run()"},{"p":"com.elangsegara","c":"XrayExtension","l":"setAction(String)","u":"setAction(java.lang.String)"},{"p":"com.elangsegara","c":"XrayExtension","l":"setScenario(String)","u":"setScenario(java.lang.String)"},{"p":"com.elangsegara","c":"XrayExtension","l":"XrayExtension()","u":"%3Cinit%3E()"},{"p":"com.elangsegara","c":"XrayMain","l":"XrayMain()","u":"%3Cinit%3E()"},{"p":"com.elangsegara","c":"XrayPlugin","l":"XrayPlugin()","u":"%3Cinit%3E()"},{"p":"com.elangsegara","c":"XrayTask","l":"XrayTask()","u":"%3Cinit%3E()"}];updateSearchResults();
