--- docs/package-search-index.js
@@ -1,1 +1,0 @@
-packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"com.elangsegara"}];updateSearchResults();
