--- docs/tag-search-index.js
@@ -1,1 +1,0 @@
-tagSearchIndex = [];updateSearchResults();
