--- docs/type-search-index.js
@@ -1,1 +1,0 @@
-typeSearchIndex = [{"l":"All Classes and Interfaces","u":"allclasses-index.html"},{"p":"com.elangsegara","l":"XrayExtension"},{"p":"com.elangsegara","l":"XrayMain"},{"p":"com.elangsegara","l":"XrayPlugin"},{"p":"com.elangsegara","l":"XrayTask"}];updateSearchResults();
